// api/types/paths/pet/{petId}.types.ts
export type getPetById = (
  $: OmitValueWhenNever<{
    query: never;
    path: getPetById_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Pet;
          };
          "application/xml": {
            schema: Pet;
          };
        };
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      [statusCode in Exclude<HttpStatusCode, 200 | 400 | 404>]: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
  }>,
) => MaybePromise<
  | {
      status: 200;
      contentType?: "application/json";
      body?: Pet;
    }
  | {
      status: 200;
      contentType?: "application/xml";
      body?: Pet;
    }
  | {
      status: 400;
    }
  | {
      status: 404;
    }
  | {
      status: number | undefined;
    }
  | { status: 415; contentType: "text/plain"; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE }
>;

/**
 * Updates a pet resource based on the form data.
 */
export type updatePetWithForm = (
  $: OmitValueWhenNever<{
    query: updatePetWithForm_Query;
    path: updatePetWithForm_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Pet;
          };
          "application/xml": {
            schema: Pet;
          };
        };
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      [statusCode in Exclude<HttpStatusCode, 200 | 400>]: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
  }>,
) => MaybePromise<
  | {
      status: 200;
      contentType?: "application/json";
      body?: Pet;
    }
  | {
      status: 200;
      contentType?: "application/xml";
      body?: Pet;
    }
  | {
      status: 400;
    }
  | {
      status: number | undefined;
    }
  | { status: 415; contentType: "text/plain"; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE }
>;

/**
 * Delete a pet.
 */
export type deletePet = (
  $: OmitValueWhenNever<{
    query: never;
    path: deletePet_Path;
    headers: deletePet_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      [statusCode in Exclude<HttpStatusCode, 200 | 400>]: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
  }>,
) => MaybePromise<
  | {
      status: 200;
    }
  | {
      status: 400;
    }
  | {
      status: number | undefined;
    }
  | { status: 415; contentType: "text/plain"; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE }
>;

export type getPetById_Path = {
  /**
   * ID of pet to return
   */
  petId: number;
};

export type updatePetWithForm_Query = {
  /**
   * Name of pet that needs to be updated
   */
  name?: string;
  /**
   * Status of pet that needs to be updated
   */
  status?: string;
};

export type updatePetWithForm_Path = {
  /**
   * ID of pet that needs to be updated
   */
  petId: number;
};

export type deletePet_Path = {
  /**
   * Pet id to delete
   */
  petId: number;
};

export type deletePet_Headers = { api_key?: string };